+++ airflow/operators/gcs_to_sftp.py
@@ -26,7 +26,7 @@ from typing import Optional
 from airflow import AirflowException
 from airflow.gcp.hooks.gcs import GCSHook
 from airflow.models import BaseOperator
-from airflow.providers.sftp.hooks.sftp_hook import SFTPHook
+from airflow.providers.sftp.hooks.sftp import SFTPHook
 from airflow.utils.decorators import apply_defaults
 
 WILDCARD = "*"
