+++ docs/operators-and-hooks-ref.rst
@@ -1275,7 +1275,7 @@ communication protocols or interface.
 
    * - `SSH File Transfer Protocol (SFTP) <https://tools.ietf.org/wg/secsh/draft-ietf-secsh-filexfer/>`__
      -
-     - :mod:`airflow.providers.sftp.hooks.sftp_hook`
+     - :mod:`airflow.providers.sftp.hooks.sftp`
      - :mod:`airflow.providers.sftp.operators.sftp_operator`
      - :mod:`airflow.providers.sftp.sensors.sftp_sensor`
