+++ tests/providers/sftp/hooks/test_sftp.py
@@ -26,7 +26,7 @@ import pysftp
 from parameterized import parameterized
 
 from airflow.models import Connection
-from airflow.providers.sftp.hooks.sftp_hook import SFTPHook
+from airflow.providers.sftp.hooks.sftp import SFTPHook
 from airflow.utils.session import provide_session
 
 TMP_PATH = '/tmp'
@@ -145,14 +145,14 @@ class TestSFTPHook(unittest.TestCase):
             TMP_PATH, TMP_DIR_FOR_TESTS, TMP_FILE_FOR_TESTS))
         self.assertEqual(len(output), 14)
 
-    @mock.patch('airflow.providers.sftp.hooks.sftp_hook.SFTPHook.get_connection')
+    @mock.patch('airflow.providers.sftp.hooks.sftp.SFTPHook.get_connection')
     def test_no_host_key_check_default(self, get_connection):
         connection = Connection(login='login', host='host')
         get_connection.return_value = connection
         hook = SFTPHook()
         self.assertEqual(hook.no_host_key_check, False)
 
-    @mock.patch('airflow.providers.sftp.hooks.sftp_hook.SFTPHook.get_connection')
+    @mock.patch('airflow.providers.sftp.hooks.sftp.SFTPHook.get_connection')
     def test_no_host_key_check_enabled(self, get_connection):
         connection = Connection(
             login='login', host='host',
@@ -162,7 +162,7 @@ class TestSFTPHook(unittest.TestCase):
         hook = SFTPHook()
         self.assertEqual(hook.no_host_key_check, True)
 
-    @mock.patch('airflow.providers.sftp.hooks.sftp_hook.SFTPHook.get_connection')
+    @mock.patch('airflow.providers.sftp.hooks.sftp.SFTPHook.get_connection')
     def test_no_host_key_check_disabled(self, get_connection):
         connection = Connection(
             login='login', host='host',
@@ -172,7 +172,7 @@ class TestSFTPHook(unittest.TestCase):
         hook = SFTPHook()
         self.assertEqual(hook.no_host_key_check, False)
 
-    @mock.patch('airflow.providers.sftp.hooks.sftp_hook.SFTPHook.get_connection')
+    @mock.patch('airflow.providers.sftp.hooks.sftp.SFTPHook.get_connection')
     def test_no_host_key_check_disabled_for_all_but_true(self, get_connection):
         connection = Connection(
             login='login', host='host',
@@ -182,7 +182,7 @@ class TestSFTPHook(unittest.TestCase):
         hook = SFTPHook()
         self.assertEqual(hook.no_host_key_check, False)
 
-    @mock.patch('airflow.providers.sftp.hooks.sftp_hook.SFTPHook.get_connection')
+    @mock.patch('airflow.providers.sftp.hooks.sftp.SFTPHook.get_connection')
     def test_no_host_key_check_ignore(self, get_connection):
         connection = Connection(
             login='login', host='host',
@@ -192,7 +192,7 @@ class TestSFTPHook(unittest.TestCase):
         hook = SFTPHook()
         self.assertEqual(hook.no_host_key_check, True)
 
-    @mock.patch('airflow.providers.sftp.hooks.sftp_hook.SFTPHook.get_connection')
+    @mock.patch('airflow.providers.sftp.hooks.sftp.SFTPHook.get_connection')
     def test_no_host_key_check_no_ignore(self, get_connection):
         connection = Connection(
             login='login', host='host',
